+++ lib/puppet_check.rb
@@ -143,7 +143,7 @@ class PuppetCheck
     eyamls, files = files.partition { |file| File.extname(file) =~ /\.eya?ml$/ }
     DataParser.eyaml(eyamls, public, private) unless eyamls.empty?
     # check misc ruby
-    librarians, files = files.partition { |file| File.basename(file) =~ /(?:Puppet|Module|Rake|Gem)file$/ }
+    librarians, files = files.partition { |file| File.basename(file) =~ /(?:Puppet|Module|Rake|Gem|Vagrant)file|.gemspec$/ }
     RubyParser.librarian(librarians, style, rubocop_args) unless librarians.empty?
     # ignore everything else
     files.each { |file| self.class.files[:ignored].push(file.to_s) }